test(components): split tests at objects-group & objects-group-dialog
Jira issue: VEGA-3131
